@@ -1,10 +1,10 @@
 {
   "name": "react-photoswipe",
-  "version": "0.5.1",
+  "version": "1.0.0",
   "description": "PhotoSwipe, PhotoSwipeGallery component for ReactJS base on PhotoSwipe.",
   "main": "lib/index.js",
   "scripts": {
-    "start": "NODE_ENV=development node example/webpack.dev.server.js",
+    "start": "NODE_ENV=development gulp start",
     "build": "NODE_ENV=production gulp build"
   },
   "keywords": [
@@ -24,44 +24,54 @@
     "photoswipe": "^4.1.0"
   },
   "devDependencies": {
-    "autoprefixer": "^6.0.3",
-    "babel-core": "^5.8.25",
+    "autoprefixer": "^6.3.1",
+    "babel": "^5.8.34",
+    "babel-core": "^5.8.29",
     "babel-eslint": "^4.1.3",
     "babel-loader": "^5.3.2",
-    "camelcase": "^1.2.1",
-    "css-loader": "^0.19.0",
-    "del": "^2.0.2",
-    "eslint": "^1.6.0",
-    "eslint-loader": "^1.0.0",
-    "eslint-plugin-react": "^3.5.1",
+    "babel-plugin-react-transform": "^1.1.1",
+    "bootstrap": "^3.3.6",
+    "camelcase": "^2.0.1",
+    "css-loader": "^0.23.0",
+    "cssnano": "^3.3.2",
+    "del": "^2.1.0",
+    "eslint": "^1.10.1",
+    "eslint-loader": "^1.2.0",
+    "eslint-plugin-react": "^3.15.0",
     "express": "^4.13.3",
-    "file-loader": "^0.8.4",
+    "extract-text-webpack-plugin": "^1.0.1",
+    "file-loader": "^0.8.5",
     "gulp": "^3.9.0",
-    "gulp-autoprefixer": "^3.0.2",
+    "gulp-autoprefixer": "^3.1.0",
     "gulp-babel": "^5.2.1",
     "gulp-concat": "^2.6.0",
-    "gulp-eslint": "^1.0.0",
+    "gulp-eslint": "^1.1.0",
     "gulp-filter": "^3.0.1",
-    "gulp-frep": "^0.1.3",
-    "gulp-minify-css": "^1.2.1",
-    "gulp-minify-html": "^1.0.4",
+    "gulp-minify-css": "^1.2.4",
+    "gulp-postcss": "^6.1.0",
     "gulp-replace": "^0.5.4",
-    "gulp-sass": "^2.0.4",
-    "http-proxy": "^1.11.2",
-    "json-loader": "^0.5.3",
+    "gulp-sass": "^2.2.0",
+    "history": "^2.0.1",
+    "html-loader": "^0.4.0",
+    "html-webpack-plugin": "^2.7.1",
+    "http-proxy": "^1.12.0",
+    "json-loader": "^0.5.4",
     "less": "^2.5.3",
     "less-loader": "^2.2.1",
-    "lodash.merge": "^3.3.2",
-    "node-sass": "^3.3.3",
-    "opn": "^3.0.2",
-    "postcss-loader": "^0.6.0",
-    "react-hot-loader": "^1.3.0",
-    "run-sequence": "^1.1.4",
-    "sass-loader": "^3.0.0",
-    "style-loader": "^0.12.4",
-    "webpack": "^1.12.2",
-    "webpack-dev-server": "^1.12.0",
-    "webpack-stream": "^2.1.1"
+    "node-sass": "^3.4.2",
+    "opn": "^4.0.0",
+    "postcss-loader": "^0.8.0",
+    "react-router": "^2.0.0",
+    "react-transform-hmr": "^1.0.1",
+    "run-sequence": "^1.1.5",
+    "sass-loader": "^3.1.2",
+    "style-loader": "^0.13.0",
+    "task-dir": "^0.2.0",
+    "webpack": "^1.12.11",
+    "webpack-dev-middleware": "^1.2.0",
+    "webpack-hot-middleware": "^2.5.0",
+    "webpack-manifest-plugin": "^1.0.0",
+    "webpack-stream": "^3.0.1"
   },
   "repository": {
     "type": "git",
