@@ -1,26 +1,29 @@
 import React from 'react';
+import {Link} from 'react-router';
 
-const logoImg = require('../assets/images/logo.svg');
+import logo from 'assets/images/logo.svg';
 
-export default class Header extends React.Component {
+class Header extends React.Component {
   render() {
     return (
-      <header className='layout-header'>
-        <nav className='navbar navbar-inverse'>
-          <div className='container'>
-            <div className='navbar-header'>
-              <button type='button' className='navbar-toggle collapsed'>
-                <span className='sr-only'>Toggle navigation</span>
-                <span className='icon-bar'></span>
-                <span className='icon-bar'></span>
-                <span className='icon-bar'></span>
+      <header className="layout-header">
+        <nav className="navbar navbar-inverse">
+          <div className="container">
+            <div className="navbar-header">
+              <button type="button" className="navbar-toggle collapsed">
+                <span className="sr-only">Toggle navigation</span>
+                <span className="icon-bar"/>
+                <span className="icon-bar"/>
+                <span className="icon-bar"/>
               </button>
-              <a className='navbar-brand' href='#'>
-                <img width='20' src={logoImg}></img>
-              </a>
+              <Link className="navbar-brand" to="/">
+                <img width="20" src={logo}/>
+              </Link>
             </div>
-            <div className='collapse navbar-collapse'>
-              <ul className='nav navbar-nav navbar-right'></ul>
+            <div className="collapse navbar-collapse">
+              <ul className="nav navbar-nav navbar-right">
+                <li><Link to="/gallery">Gallery</Link></li>
+              </ul>
             </div>
           </div>
         </nav>
@@ -28,3 +31,6 @@ export default class Header extends React.Component {
     );
   }
 }
+
+export default Header;
+
