@@ -1,13 +1,16 @@
 import React from 'react';
 
-export default class Footer extends React.Component {
+class Footer extends React.Component {
   render() {
     return (
-      <footer className='layout-footer'>
-        <div className='container'>
-          <div className='text-center'>&copy; 2015 Lorem ipsum.</div>
+      <footer className="layout-footer">
+        <div className="container">
+          <div className="text-center">&copy; 2015 Minh Tran.</div>
         </div>
       </footer>
     );
   }
 }
+
+export default Footer;
+
