@@ -4,11 +4,17 @@ Object.defineProperty(exports, '__esModule', {
   value: true
 });
 
+var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
+
 var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
 
-var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; desc = parent = undefined; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
+var _get = function get(_x2, _x3, _x4) { var _again = true; _function: while (_again) { var object = _x2, property = _x3, receiver = _x4; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x2 = parent; _x3 = property; _x4 = receiver; _again = true; desc = parent = undefined; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+
+function _toConsumableArray(arr) { if (Array.isArray(arr)) { for (var i = 0, arr2 = Array(arr.length); i < arr.length; i++) arr2[i] = arr[i]; return arr2; } else { return Array.from(arr); } }
+
+function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in obj) { if (keys.indexOf(i) >= 0) continue; if (!Object.prototype.hasOwnProperty.call(obj, i)) continue; target[i] = obj[i]; } return target; }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
 
@@ -30,6 +36,14 @@ var _photoswipeDistPhotoswipeUiDefault = require('photoswipe/dist/photoswipe-ui-
 
 var _photoswipeDistPhotoswipeUiDefault2 = _interopRequireDefault(_photoswipeDistPhotoswipeUiDefault);
 
+var _classnames = require('classnames');
+
+var _classnames2 = _interopRequireDefault(_classnames);
+
+var _events = require('./events');
+
+var _events2 = _interopRequireDefault(_events);
+
 var PhotoSwipe = (function (_React$Component) {
   _inherits(PhotoSwipe, _React$Component);
 
@@ -45,38 +59,57 @@ var PhotoSwipe = (function (_React$Component) {
     };
 
     this.componentDidMount = function () {
-      if (_this.props.isOpen) {
-        _this.openPhotoSwipe(_this.props.items, _this.props.options);
+      var isOpen = _this.state.isOpen;
+
+      if (isOpen) {
+        _this.openPhotoSwipe(_this.props);
       }
     };
 
     this.componentWillReceiveProps = function (nextProps) {
+      var isOpen = _this.state.isOpen;
+
       if (nextProps.isOpen) {
-        if (!_this.state.isOpen) {
-          _this.openPhotoSwipe(nextProps.items, nextProps.options);
+        if (!isOpen) {
+          _this.openPhotoSwipe(nextProps);
         } else {
-          _this.photoSwipe.items.length = 0;
-          nextProps.items.forEach(function (item) {
-            _this.photoSwipe.items.push(item);
-          });
-          _this.photoSwipe.invalidateCurrItems();
-          _this.photoSwipe.updateSize(true);
+          _this.updateItems(nextProps.items);
         }
-      } else if (_this.state.isOpen) {
-        _this.photoSwipe.close();
+      } else if (isOpen) {
+        _this.closePhotoSwipe();
       }
     };
 
     this.componentWillUnmount = function () {
-      if (_this.photoSwipe) {
-        _this.photoSwipe.close();
-      }
+      _this.closePhotoSwipe();
     };
 
-    this.openPhotoSwipe = function (items, options) {
+    this.openPhotoSwipe = function (props) {
+      var items = props.items;
+      var options = props.options;
+
+      var other = _objectWithoutProperties(props, ['items', 'options']);
+
       var pswpElement = _reactDom2['default'].findDOMNode(_this);
       _this.photoSwipe = new _photoswipe2['default'](pswpElement, _photoswipeDistPhotoswipeUiDefault2['default'], items, options);
-      _this.photoSwipe.listen('close', _this.handleClose);
+      _events2['default'].forEach(function (event) {
+        var callback = props[event];
+        if (callback || event === 'destroy') {
+          (function () {
+            var self = _this;
+            _this.photoSwipe.listen(event, function () {
+              if (callback) {
+                var args = arguments.length === 1 ? [arguments[0]] : Array.apply(null, arguments);
+                args.unshift(this);
+                callback.apply(undefined, _toConsumableArray(args));
+              }
+              if (event === 'destroy') {
+                self.handleClose();
+              }
+            });
+          })();
+        }
+      });
       _this.setState({
         isOpen: true
       }, function () {
@@ -84,12 +117,32 @@ var PhotoSwipe = (function (_React$Component) {
       });
     };
 
+    this.updateItems = function () {
+      var items = arguments.length <= 0 || arguments[0] === undefined ? [] : arguments[0];
+
+      _this.photoSwipe.items.length = 0;
+      items.forEach(function (item) {
+        _this.photoSwipe.items.push(item);
+      });
+      _this.photoSwipe.invalidateCurrItems();
+      _this.photoSwipe.updateSize(true);
+    };
+
+    this.closePhotoSwipe = function () {
+      if (!_this.photoSwipe) {
+        return;
+      }
+      _this.photoSwipe.close();
+    };
+
     this.handleClose = function () {
+      var onClose = _this.props.onClose;
+
       _this.setState({
         isOpen: false
       }, function () {
-        if (_this.props.onClose) {
-          _this.props.onClose();
+        if (onClose) {
+          onClose();
         }
       });
     };
@@ -98,9 +151,29 @@ var PhotoSwipe = (function (_React$Component) {
   _createClass(PhotoSwipe, [{
     key: 'render',
     value: function render() {
+      var _props = this.props;
+      var isOpen = _props.isOpen;
+      var items = _props.items;
+      var options = _props.options;
+      var onClose = _props.onClose;
+      var className = _props.className;
+
+      var other = _objectWithoutProperties(_props, ['isOpen', 'items', 'options', 'onClose', 'className']);
+
+      className = (0, _classnames2['default'])(['pswp', className]).trim();
+      var otherProps = {};
+      for (var propName in other) {
+        if (other.hasOwnProperty(propName) && _events2['default'].indexOf(propName) === -1) {
+          otherProps[propName] = other[propName];
+        }
+      }
       return _react2['default'].createElement(
         'div',
-        { className: 'pswp', tabIndex: '-1', role: 'dialog', 'aria-hidden': 'true' },
+        _extends({}, otherProps, {
+          className: className,
+          tabIndex: '-1',
+          role: 'dialog',
+          'aria-hidden': 'true' }),
         _react2['default'].createElement('div', { className: 'pswp__bg' }),
         _react2['default'].createElement(
           'div',
@@ -166,7 +239,8 @@ var PhotoSwipe = (function (_React$Component) {
       isOpen: _react2['default'].PropTypes.bool.isRequired,
       items: _react2['default'].PropTypes.array.isRequired,
       options: _react2['default'].PropTypes.object,
-      onClose: _react2['default'].PropTypes.func
+      onClose: _react2['default'].PropTypes.func,
+      className: _react2['default'].PropTypes.string
     },
     enumerable: true
   }, {
