@@ -4,11 +4,15 @@ Object.defineProperty(exports, '__esModule', {
   value: true
 });
 
+var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
+
 var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
 
 var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; desc = parent = undefined; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+
+function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in obj) { if (keys.indexOf(i) >= 0) continue; if (!Object.prototype.hasOwnProperty.call(obj, i)) continue; target[i] = obj[i]; } return target; }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
 
@@ -25,6 +29,14 @@ var _reactDom2 = _interopRequireDefault(_reactDom);
 var _PhotoSwipeJs = require('./PhotoSwipe.js');
 
 var _PhotoSwipeJs2 = _interopRequireDefault(_PhotoSwipeJs);
+
+var _events = require('./events');
+
+var _events2 = _interopRequireDefault(_events);
+
+var _classnames = require('classnames');
+
+var _classnames2 = _interopRequireDefault(_classnames);
 
 var PhotoSwipeGallery = (function (_React$Component) {
   _inherits(PhotoSwipeGallery, _React$Component);
@@ -45,6 +57,7 @@ var PhotoSwipeGallery = (function (_React$Component) {
       return function (e) {
         e.preventDefault();
         var options = _this.state.options;
+
         options.index = itemIndex;
         options.getThumbBoundsFn = function (index) {
           var thumbnail = _reactDom2['default'].findDOMNode(_this.refs['thumbnail' + index]);
@@ -72,24 +85,50 @@ var PhotoSwipeGallery = (function (_React$Component) {
     value: function render() {
       var _this2 = this;
 
+      var _props = this.props;
+      var className = _props.className;
+      var items = _props.items;
+      var thumbnailContent = _props.thumbnailContent;
+
+      var other = _objectWithoutProperties(_props, ['className', 'items', 'thumbnailContent']);
+
+      var _state = this.state;
+      var isOpen = _state.isOpen;
+      var options = _state.options;
+
+      className = (0, _classnames2['default'])(['pswp-gallery', className]).trim();
+      var otherProps = {};
+      var eventProps = {};
+      for (var propName in other) {
+        if (other.hasOwnProperty(propName)) {
+          if (_events2['default'].indexOf(propName) === -1) {
+            otherProps[propName] = other[propName];
+          } else {
+            eventProps[propName] = other[propName];
+          }
+        }
+      }
       return _react2['default'].createElement(
         'div',
-        { className: 'pswp-gallery' },
+        _extends({}, otherProps, { className: className }),
         _react2['default'].createElement(
           'div',
           { className: 'pswp-thumbnails' },
-          this.props.items.map(function (item, index) {
+          items.map(function (item, index) {
             return _react2['default'].createElement(
               'div',
               { key: index, ref: 'thumbnail' + index,
                 className: 'pswp-thumbnail',
                 onClick: _this2.showPhotoSwipe(index) },
-              _this2.props.thumbnailContent(item)
+              thumbnailContent(item)
             );
           })
         ),
-        _react2['default'].createElement(_PhotoSwipeJs2['default'], { isOpen: this.state.isOpen, items: this.props.items,
-          options: this.state.options, onClose: this.handleClose })
+        _react2['default'].createElement(_PhotoSwipeJs2['default'], _extends({}, eventProps, {
+          isOpen: isOpen,
+          items: items,
+          options: options,
+          onClose: this.handleClose }))
       );
     }
   }], [{
@@ -97,7 +136,8 @@ var PhotoSwipeGallery = (function (_React$Component) {
     value: {
       items: _react2['default'].PropTypes.array.isRequired,
       options: _react2['default'].PropTypes.object,
-      thumbnailContent: _react2['default'].PropTypes.func
+      thumbnailContent: _react2['default'].PropTypes.func,
+      className: _react2['default'].PropTypes.string
     },
     enumerable: true
   }, {
